Upgrade to Wicket 6.11.0
git-svn-id: https://svn.apache.org/repos/asf/syncope/branches/1_1_X@1528104 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/syncope/core/workflow/user/activiti/ActivitiUserWorkflowAdapter.java
@@ -365,7 +365,11 @@ public class ActivitiUserWorkflowAdapter extends AbstractUserWorkflowAdapter {
         doExecuteTask(user, "delete", null);
         userDAO.delete(user);
 
-        historyService.deleteHistoricProcessInstance(user.getWorkflowId());
+        if (!historyService.createHistoricProcessInstanceQuery().
+                processInstanceId(user.getWorkflowId()).list().isEmpty()) {
+
+            historyService.deleteHistoricProcessInstance(user.getWorkflowId());
+        }
     }
 
     @Override
